requirements.txt
streamlit>=1.33
pandas>=2.0
numpy>=1.26
requests>=2.31
python-dotenv>=1.0
pillow>=10.0
